requirements.txt
streamlit
speechrecognition
openai
Pillow
